Restructure so it's more package-like
