+++ godwin/__init__.py
@@ -1,0 +1,2 @@
+from .Database import Database
+from .Scraper import Scraper
